Task 4: Extract agent/session_pickup.py — pop locking, startup steering drain, dependency checks
